+++ src/test/kotlin/WinnerTest.kt
@@ -1,5 +1,5 @@
 import data.Car
-import domain.game.CarRacingGame
+import domain.CarRacingGame
 import domain.generator.NumberGenerator
 import domain.generator.RacingNumberGenerator
 import org.assertj.core.api.Java6Assertions.assertThat
